Refactor app layout
