+++ src/styles/index.css
@@ -3,3 +3,7 @@ body {
   padding: 0;
   font-family: sans-serif;
 }
+
+body, .App, #router, #root {
+  height: 100%;
+}
